+++ src/components/dashboard/SeriesDetailsDashboard.jsx
@@ -1,23 +1,27 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import Swal from 'sweetalert2'; // Importa SweetAlert2
+
 function SeriesDetailsDashboard({ seriesId }) {
   const [seriesDetails, setSeriesDetails] = useState(null);
-
- const fetchSeriesDetails = async () => {
-      try {
-        const response = await fetch(`https://duwi.onrender.com/series/${seriesId}`);
-        if (response.status === 200) {
-          const seriesDetailsData = await response.json();
-          setSeriesDetails(seriesDetailsData);
-        }
-      } catch (error) {
-        console.error('Error al obtener detalles de la serie:', error);
+  const [showCard, setShowCard] = useState(true)
+  const fetchSeriesDetails = async () => {
+    try {
+      const response = await fetch(`https://duwi.onrender.com/series/${seriesId}`);
+      if (response.status === 200) {
+        const seriesDetailsData = await response.json();
+        setSeriesDetails(seriesDetailsData);
       }
-    };
+    } catch (error) {
+      console.error('Error al obtener detalles de la serie:', error);
+    }
+  };
+
+  
 
   useEffect(() => {
     fetchSeriesDetails();
-  }, [seriesId,seriesDetails]);
+  }, [seriesId, seriesDetails]);
 
   const userData = localStorage.getItem('userData');
   const userInfo = JSON.parse(userData);
@@ -25,13 +29,25 @@ function SeriesDetailsDashboard({ seriesId }) {
   const handleRemoveFromFavorites = async () => {
     const userId = userInfo._id;
 
-    try {
-     const response= await axios.delete(`/favorites/${userId}/${seriesId}`);
-    } catch (error) {
-      console.error('Error al eliminar película de favoritos:', error);
-    }
+    // Muestra un mensaje de confirmación con SweetAlert2
+    Swal.fire({
+      title: `¿Estás seguro de eliminar "${seriesDetails.name}" de tus series favoritas?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        // El usuario confirmó la eliminación, realiza la eliminación
+     
+         await axios.delete(`/favorites/${userId}/${seriesId}`);
+         setShowCard(false)
+       
+      }
+    });
   };
-  return seriesDetails ? (
+
+  return  showCard && seriesDetails ? (
     <div className="w-48 p-2 rounded-lg shadow-md mb-4 mr-5">
       <div className="relative rounded-lg overflow-hidden border-t-0">
         <img
@@ -42,7 +58,9 @@ function SeriesDetailsDashboard({ seriesId }) {
       </div>
       <div className="mt-2">
         <h3 className="text-lg font-medium font-poppins text-moradito dark:text-clarito">{seriesDetails.name}</h3>
-        <button className="text-lila hover:text-moradito" onClick={handleRemoveFromFavorites}>Eliminar de Favoritos</button>
+        <button className="text-lila hover:text-moradito" onClick={handleRemoveFromFavorites}>
+          Eliminar de Favoritos
+        </button>
       </div>
     </div>
   ) : null;
